Finished Sandbox Project
Thanks to the help from Mr. G I completed this mini project.
Note:
I failed to realize how the querySelectorAll works. It's not an array creator, it creates objects. which are key value pairs.
so in order to properly add up the values of the inputs we need to iterate through every "key" adding their values together one by one.